@@ -1,11 +1,24 @@
 cmake_minimum_required(VERSION 3.22)
 project(serial)
 
-set(CMAKE_CXX_STANDARD 17)
+
 SET(IS_STATIC "STATIC" CACHE STRING "If true compile static")
 SET(WITH_DEMO "DEMO" CACHE STRING "If true compile the demo")
 
+if(NOT CMAKE_BUILD_TYPE)
+set(CMAKE_BUILD_TYPE Release)
+endif()
+set(CMAKE_CXX_FLAGS_RELEASE -Ofast)
+
+# # 设置 C++ 标准
+set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_STANDARD_REQUIRED True)
+
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fno-exceptions -fno-rtti")
+
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} ${CMAKE_CURRENT_SOURCE_DIR}/cmake)
+
+
 include_directories(include)
 file(GLOB LIB_SOURCES
     "include/ofSerial.h"
